Refactor punishments to seperate Server, Website and Discord (#54)
* Update dbinit.sql Split punishments table into webPunishments and discordPunishments

* Update Schema

Change punishmentId column of appeals table to a TEXT field so that Id's can be stored as S-1, W-1, D-1. Where S = server, W = web, and D = Discord.

Removed the appealed column from both punishment tables. This column will only exist in the punishments VIEW (added in future commit) and it will be determined based on whether or not an appeal exists for that punishment.

* Add punishments VIEW to database

Co-authored-by: Aron Brown <webmaster@birdwingfx.com>